requirements.txt
amqp==1.4.9
anyjson==0.3.3
attrs==16.2.0
billiard==3.3.0.23
celery==3.1.24
coverage==4.2
Django==1.10.2
django-filter==0.15.2
djangorestframework==3.4.7
kombu==3.0.36
Markdown==2.6.7
nose==1.3.7
psycopg2==2.6.2
pymongo==3.3.0
python-dateutil==2.5.3
pytz==2016.7
PyYAML==3.12
requests==2.11.1
six==1.10.0
tqdm==4.8.4
vcrpy==1.10.3
wrapt==1.10.8

# pynder from github
-e git+git@github.com:charliewolf/pynder.git#egg=pynder